# docs/openapi-v2.yaml
openapi: 3.0.3
info:
  title: Dental PMS/CRM API v2
  version: 0.1.0
  description: |
    PMS/CRM extension for the dental-api. v2 routes layered on top of v1
    (which is frozen — see tests/test_contract_v1.py). All routes are
    multi-tenant via X-Clinic-Id (default: "default"). All non-auth
    routes require Bearer JWT when V2_REQUIRE_AUTH=true.

    Tracks may extend this spec by adding new paths. Do not change a
    schema once a track has shipped against it.
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://dental-api.example.run.app
    description: Cloud Run
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XClinicId:
      name: X-Clinic-Id
      in: header
      required: false
      schema: { type: string, default: default }
  schemas:
...
    LeadOut:
      type: object
      required: [id, clinic_id, phone]
      properties:
        id: { type: string }
        clinic_id: { type: string }
        name: { type: string, nullable: true }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
        phone: { type: string }
        email: { type: string, nullable: true }
        source: { type: string, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        owner_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LeadActivity:
      type: object
      required: [id, lead_id, kind, occurred_at]
      properties:
        id: { type: string }
        lead_id: { type: string }
        kind: { type: string, enum: [note, call, email, meeting] }
        body: { type: string, nullable: true }
        occurred_at: { type: string, format: date-time }
        payload: { type: object, nullable: true }
...
security:
  - bearerAuth: []
  - {}
paths:
...
  /api/v2/crm/leads/{id}/convert:
    parameters:
      - { $ref: "#/components/parameters/XClinicId" }
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      summary: Convert a lead to a patient (idempotent)
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { create_patient: { type: boolean, default: true } } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, required: [lead_id, patient_id], properties: { lead_id: { type: string }, patient_id: { type: string }, status: { type: string } } } } } }
...
  /api/v2/crm/leads:
    parameters: [{ $ref: "#/components/parameters/XClinicId" }]
    post:
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [first_name, last_name, phone]
              properties:
                first_name: { type: string }
                last_name: { type: string }
                phone: { type: string }
                email: { type: string, nullable: true }
                source: { type: string, nullable: true }
                notes: { type: string, nullable: true }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/LeadOut" } } } }
  /api/v2/crm/leads/{lead_id}:
    parameters:
      - { $ref: "#/components/parameters/XClinicId" }
      - { name: lead_id, in: path, required: true, schema: { type: string } }
    put:
      summary: Partial update a lead (owner_id, status, notes, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_id: { type: string, nullable: true }
                status: { type: string, nullable: true }
                notes: { type: string, nullable: true }
                first_name: { type: string, nullable: true }
                last_name: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                email: { type: string, nullable: true }
                source: { type: string, nullable: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/LeadOut" } } } }
        '404': { description: Lead not found }
  /api/v2/crm/leads/{lead_id}/activities:
    parameters:
      - { $ref: "#/components/parameters/XClinicId" }
      - { name: lead_id, in: path, required: true, schema: { type: string } }
    post:
      summary: Create an activity (note/call/email/meeting) for a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, body]
              properties:
                kind: { type: string, enum: [note, call, email, meeting] }
                body: { type: string }
                payload: { type: object, nullable: true }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/LeadActivity" } } } }
        '404': { description: Lead not found }
    get:
      summary: List activities for a lead, sorted by created_at DESC
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/LeadActivity" } } } } }
        '404': { description: Lead not found }
...